Update blasebot.py
This is a fully functioning Windows version for blasebot that will max bet on the higher win percentage team. This version is capable of being run 24hours however, as mentioned in the code comments; it is HIGHLY recommended that you manually zoom your automated web browser to 30% upon launching for the first time. 

This code will function on most Linux distributions if you manually convert the email/password paths to Linux appropriate paths. 

A Raspberry Pi ARMv7 version is also in the works but due to the stability of browsers on the Pi and lack of web driver compatibility, this has proven to be a challenging task.